# Improve calculating symmetric difference