--- a/pkg/controller/networkpolicy/networkpolicy_controller.go
+++ b/pkg/controller/networkpolicy/networkpolicy_controller.go
@@ -984,7 +984,7 @@ func (n *NetworkPolicyController) updatePod(oldObj, curObj interface{}) {
 	} else if !labelsEqual {
 		// No need to enqueue common AppliedToGroups as they already have latest Pod
 		// information.
-		appliedToGroupKeys = oldAppliedToGroupKeySet.Difference(curAppliedToGroupKeySet).Union(curAppliedToGroupKeySet.Difference(oldAppliedToGroupKeySet))
+		appliedToGroupKeys = utilsets.SymmetricDifference(oldAppliedToGroupKeySet, curAppliedToGroupKeySet)
 	}
 	// AddressGroup keys must be enqueued only if the Pod's IP has changed or
 	// if Pod's label change causes it to match new Groups. Same applies for ClusterGroups.
@@ -994,8 +994,8 @@ func (n *NetworkPolicyController) updatePod(oldObj, curObj interface{}) {
 	} else if !labelsEqual {
 		// No need to enqueue common AddressGroups as they already have latest Pod
 		// information.
-		addressGroupKeys = oldAddressGroupKeySet.Difference(curAddressGroupKeySet).Union(curAddressGroupKeySet.Difference(oldAddressGroupKeySet))
-		groupKeys = oldGroupKeySet.Difference(curGroupKeySet).Union(curGroupKeySet.Difference(oldGroupKeySet))
+		addressGroupKeys = utilsets.SymmetricDifference(oldAddressGroupKeySet, curAddressGroupKeySet)
+		groupKeys = utilsets.SymmetricDifference(oldGroupKeySet, curGroupKeySet)
 	}
 	for group := range appliedToGroupKeys {
 		n.enqueueAppliedToGroup(group)
@@ -1082,11 +1082,11 @@ func (n *NetworkPolicyController) updateNamespace(oldObj, curObj interface{}) {
 	oldGroupKeySet := n.filterInternalGroupsForNamespace(oldNamespace)
 	// No need to enqueue common AddressGroups as they already have latest
 	// Namespace information.
-	addressGroupKeys := oldAddressGroupKeySet.Difference(curAddressGroupKeySet).Union(curAddressGroupKeySet.Difference(oldAddressGroupKeySet))
+	addressGroupKeys := utilsets.SymmetricDifference(oldAddressGroupKeySet, curAddressGroupKeySet)
 	for group := range addressGroupKeys {
 		n.enqueueAddressGroup(group)
 	}
-	groupKeys := oldGroupKeySet.Difference(curGroupKeySet).Union(curGroupKeySet.Difference(oldGroupKeySet))
+	groupKeys := utilsets.SymmetricDifference(oldGroupKeySet, curGroupKeySet)
 	for group := range groupKeys {
 		n.enqueueInternalGroup(group)
 	}