Calculating symmetric difference is common in the code, while
s1.Difference(s2).Union(s2.Difference(s1)) is a little complicated,
error-prone, and always builds several unnecessary intermediate sets.
This patch adds an util function SymmetricDifference to improve it. The
benchmark impact is as below when calculating symmetric difference of
two sets each of which contains 2000 items:

```
name                    old time/op    new time/op    delta
SymmetricDifference-48    1.50ms ± 1%    0.87ms ± 3%  -42.10%  (p=0.008 n=5+5)

name                    old alloc/op   new alloc/op   delta
SymmetricDifference-48     342kB ± 0%     171kB ± 0%  -49.99%  (p=0.008 n=5+5)

name                    old allocs/op  new allocs/op  delta
SymmetricDifference-48       132 ± 0%        61 ± 0%  -53.79%  (p=0.029 n=4+4)
```